LocalCustoms

Documentation for LocalCustoms.

LocalCustoms.@local_

LocalCustoms.@localexport

LocalCustoms.@localimport

LocalCustoms.@local_ — Macro

@local_ import ModuleName
@local_ import ModuleName: name
@local_ import ModuleName: name1, name2, ...

Import names belonging to a local module (i.e. already in the namespace).

Do not include a . before the module name!

When only a composite ModuleName is provided (e.g. Gadgets.Things.Doohickies), the innermost module (i.e. Doohickies) is imported into the active namespace.

Otherwise, this macro is just a verbose way of avoiding the .. :)

@local_ export ModuleName
@local_ export ModuleName: name
@local_ export ModuleName: name1, name2, ...

Export names belonging to a local module (i.e. already in the namespace).

Do not include a . before the module name!

The point of this macro is to export names of a subordinate module without having to (explicitly) import them first.

When a nested module is provided (e.g. Gadgets.Things.Doohickies), ...this macro doesn't work. Sorry. The Julia parser doesn't expect any operators after an export name. It's a total fluke that the : signatures parse as expressions at all. You can use the alternative macro @localexport for such cases.

LocalCustoms.@localexport — Macro

@localexport ModuleName
@localexport ModuleName name1
@localexport ModuleName name1 name2 ...

Export names belonging to a local module (i.e. already in the namespace).

Do not include a . before the module name!

The chief difference between this macro and @local_ export is that this macro supports a composite ModuleName like Gadgets.Things.Doohickies.

When only a composite ModuleName is provided (e.g. Gadgets.Things.Doohickies), the innermost module (i.e. Doohickies) is exported from the active namespace.

LocalCustoms.@localimport — Macro

@localimport ModuleName
@localimport ModuleName name1
@localimport ModuleName name1 name2 ...

Import names belonging to a local module (i.e. already in the namespace).

Do not include a . before the module name!

The only difference between this macro and @local_ import is the syntax. Both are provided for the sake of symmetry with exports.

When only a composite ModuleName is provided (e.g. Gadgets.Things.Doohickies), the innermost module (i.e. Doohickies) is imported into the active namespace.